harder boss1

+++ AnythingButFixed.py
@@ -6,16 +6,15 @@ Update notes:
 -Bill
 -faster
 -shorter
--now without sys
 -added bats
 -now with sys
--added a list of shareholders
 -3 new bosses
 -4 new weapons
 -2 new weapons
 -buffed boss 1
 -not as random
 -added Todolist
+-now without sys
 -60% less gluten
 -120% more gluten
 -changed tutorial
@@ -31,6 +30,7 @@ Update notes:
 -now using the metric system
 -Fixed hitboxes for everyone
 -now with git (#notsponsored)
+-added a list of shareholders
 -sorted update notes by length
 -maybe buffed blaster 2.0 (we forgot)
 -fixed boss 2 and also made him exist
@@ -245,7 +245,7 @@ class player(Turtle):
         return
             
 class bullet(Turtle):
-    def __init__(self, direction, pos, color = (255, 0, 0), sp = 1.5, btype = 'regular', explosion = 1, debuffs = {}):
+    def __init__(self, direction, pos, color = (255, 0, 0), sp = 1.5, btype = 'regular', explosion = 1, debuffs = {}, torus = False):
         Turtle.__init__(self)
         self.movespeed = sp #How fast you are (higher is faster)
         self.damage = 0 #How much damage a bullet deals
@@ -260,6 +260,7 @@ class bullet(Turtle):
         self.direction = direction
         self.seth(self.direction)
         self.explosion = explosion
+        self.torus = torus
 
     def start(self):
         self.up()
@@ -280,13 +281,26 @@ class bullet(Turtle):
             if bestenemy != '':
                 x = self.heading()-90
                 self.seth(x+(self.towards(bestenemy)-90>x)-(self.towards(bestenemy)-90<x)+90)
-            self.forward(self.movespeed)
-        else:
-            self.forward(self.movespeed)
+        elif self.btype == 'phoming':
+            t = self.towards(p)
+            s = t-270
+            if s > 0:
+                self.left(1)
+            else:
+                self.right(1)
+            if abs(self.heading()-270) > 30:
+                if self.heading()-270 > 0:
+                    self.seth(300)
+                else:
+                    self.seth(240)
+        self.forward(self.movespeed)
         if self.ycor() < -300 or self.ycor() > 300: #Take yourself off the screen when you're off the screen
             self.delete()
         if self.xcor() < -300 or self.xcor() > 300:
-            self.delete()
+            if self.torus:
+                self.setx(round(self.xcor()*-0.99))
+            else:
+                self.delete()
 
     def moveToPos(self, pos):
         self.goto(pos)
@@ -533,8 +547,8 @@ class Boss(Turtle):
         if self.health <= 0:
             self.delete()
                     
-    def shoot(self, direction = -90, debuffs = {}):
-        b = bullet(direction, self.pos(), debuffs = debuffs)
+    def shoot(self, direction = -90, debuffs = {}, btype = 'regular', torus = False):
+        b = bullet(direction, self.pos(), debuffs = debuffs, btype = btype, torus = torus)
         if 'freeze' in debuffs.keys():
             b.color((0, 255, 255))
         elif 'ion' in debuffs.keys():
@@ -618,10 +632,10 @@ class boss1(Boss):
             self.n *= -1
 
     def fire(self):
-        if not random.randint(0, 200):
+        if not random.randint(0, 100):
             self.fireenemy(1, 4)
-        if not random.randint(0, 200):
-            self.spray(3, 1, 2)
+        if not random.randint(0, 100):
+            self.spray(5, 1, 2)
 
 class boss2(Boss):
     def __init__(self):
@@ -632,7 +646,7 @@ class boss2(Boss):
         self.turtlesize(5, 5, 2)
         self.pencolor((255, 0, 0))
         self.goto(0, 300)
-        self.health = 99
+        self.health = 200
         self.showhealth()
         self.points = 100
         self.spraying = 0
@@ -647,10 +661,6 @@ class boss2(Boss):
             self.setx(-300)
 
     def fire(self):
-##        if abs(self.xcor()-self.spot) < 5:
-##            if not random.randint(0, 100):
-##                self.lazershot(self.pos(), -90)
-##                self.spot = p.xcor()
         alimit = self.health/10
         slimit = 100-self.health/4
         if self.health > 195:
@@ -664,14 +674,14 @@ class boss2(Boss):
                     self.spraying -= 1
                     if self.health > 100:
                         if random.randint(0, 1):
-                            self.shoot(direction = random.randint(-30, 30)-90, debuffs = {'freeze':30})
+                            self.shoot(direction = random.randint(-30, 30)-90, debuffs = {'freeze':30}, torus = True)
                         else:
-                            self.shoot(direction = random.randint(-30, 30)-90)
+                            self.shoot(direction = random.randint(-30, 30)-90, torus = True)
                     else:
                         if random.randint(0, 1):
-                            self.shoot(direction = random.randint(-30, 30)-90, debuffs = {'freeze':30}, btype = 'phoming')
+                            self.shoot(direction = random.randint(-30, 30)-90, debuffs = {'freeze':30}, btype = 'phoming', torus = True)
                         else:
-                            self.shoot(direction = random.randint(-30, 30)-90, btype = 'phoming')
+                            self.shoot(direction = random.randint(-30, 30)-90, btype = 'phoming', torus = True)
                 self.alternate += 1
                 if self.alternate >= alimit:
                     self.alternate = 0
@@ -1135,8 +1145,8 @@ flist = [] #Holds all the friendly's
 garbage = []
 
 mov = 0
-distance = 0## 0
-kdistance = 0## 0
+distance = 990## 0
+kdistance = 19## 0
 cdistance = 0#This is the charge count
 fight = False
 stopped = False
